Split Vivado types into separate file
