+++ hls4ml/backends/fpga/fpga_types.py
@@ -234,18 +234,6 @@ class VariableDefinition:
 # region ArrayVariable
 
 
-class VivadoArrayVariableDefinition(VariableDefinition):
-    def definition_cpp(self, name_suffix='', as_reference=False):
-        return '{type} {name}{suffix}[{shape}]'.format(
-            type=self.type.name, name=self.name, suffix=name_suffix, shape=self.size_cpp()
-        )
-
-
-class VivadoInplaceArrayVariableDefinition(VariableDefinition):
-    def definition_cpp(self):
-        return f'auto& {self.name} = {self.input_var.name}'
-
-
 class ArrayVariableConverter:
     def __init__(self, type_converter, prefix, definition_cls):
         self.type_converter = type_converter
@@ -261,16 +249,6 @@ class ArrayVariableConverter:
 
         tensor_var.__class__ = type(self.prefix + 'ArrayVariable', (type(tensor_var), self.definition_cls), {})
         return tensor_var
-
-
-class VivadoArrayVariableConverter(ArrayVariableConverter):
-    def __init__(self, type_converter):
-        super().__init__(type_converter=type_converter, prefix='Vivado', definition_cls=VivadoArrayVariableDefinition)
-
-
-class VivadoInplaceArrayVariableConverter(ArrayVariableConverter):
-    def __init__(self, type_converter):
-        super().__init__(type_converter=type_converter, prefix='Vivado', definition_cls=VivadoInplaceArrayVariableDefinition)
 
 
 # endregion
@@ -305,21 +283,6 @@ class StructMemberVariableConverter:
 # region StreamVariable
 
 
-class VivadoStreamVariableDefinition(VariableDefinition):
-    def definition_cpp(self, name_suffix='', as_reference=False):
-        if as_reference:  # Function parameter
-            return f'hls::stream<{self.type.name}> &{self.name}{name_suffix}'
-        else:  # Declaration
-            return 'hls::stream<{type}> {name}{suffix}("{name}")'.format(
-                type=self.type.name, name=self.name, suffix=name_suffix
-            )
-
-
-class VivadoInplaceStreamVariableDefinition(VariableDefinition):
-    def definition_cpp(self):
-        return f'auto& {self.name} = {self.input_var.name}'
-
-
 class StreamVariableConverter:
     def __init__(self, type_converter, prefix, definition_cls):
         self.type_converter = type_converter
@@ -341,11 +304,6 @@ class StreamVariableConverter:
         return tensor_var
 
 
-class VivadoStreamVariableConverter(StreamVariableConverter):
-    def __init__(self, type_converter):
-        super().__init__(type_converter=type_converter, prefix='Vivado', definition_cls=VivadoStreamVariableDefinition)
-
-
 # endregion
 
 # region InplaceStreamVariable
@@ -363,13 +321,6 @@ class InplaceStreamVariableConverter(StreamVariableConverter):
 
         tensor_var.__class__ = type(self.prefix + 'StreamVariable', (type(tensor_var), self.definition_cls), {})
         return tensor_var
-
-
-class VivadoInplaceStreamVariableConverter(InplaceStreamVariableConverter):
-    def __init__(self, type_converter):
-        super().__init__(
-            type_converter=type_converter, prefix='Vivado', definition_cls=VivadoInplaceStreamVariableDefinition
-        )
 
 
 # endregion
